Move to use brexpressions module and some smaller updates.
