+++ Gruntfile.js
@@ -25,17 +25,8 @@ module.exports = function(grunt) {
       options: {
         jshintrc: '.jshintrc'
       },
-      lib: ['lib/**/*.js', '!lib/expression/parser.js', 'Gruntfile.js'],
+      lib: ['lib/**/*.js', 'Gruntfile.js'],
       test: ['test/**/*.js']
-    },
-    peg: {
-      expression: {
-        options: {
-          allowedStartRules: ['start', 'expression']
-        },
-        src: "lib/expression/grammar.pegjs",
-        dest: "lib/expression/parser.js"
-      }
     },
     release: {},
     browserify: {
@@ -79,6 +70,5 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-contrib-watch');
   grunt.loadNpmTasks('grunt-simple-mocha');
   grunt.loadNpmTasks('grunt-release');
-  grunt.loadNpmTasks('grunt-peg');
   grunt.loadNpmTasks('grunt-browserify');
 };
