+++ lib/context.js
@@ -1,4 +1,4 @@
-var evaluate = require('./expression/evaluator');
+var evaluate = require('./evaluator');
 
 var Context = module.exports = function (data, parent, key) {
 	this.data = data;
