+++ lib/evaluator.js
@@ -1,4 +1,4 @@
-var Parser = require('./parser');
+var Parser = require('brexpressions');
 var textCache = {};
 var expressionCache = {};
