+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/daffl/breezy",
   "dependencies": {
+    "brexpressions": "^0.1.0",
     "data-set": "^4.0.0",
     "htmlparser": "^1.7.7",
     "virtual-dom": "^1.0.0"
@@ -32,7 +33,6 @@
     "grunt-cli": "^0.1.13",
     "grunt-contrib-jshint": "^0.10.0",
     "grunt-contrib-watch": "^0.6.1",
-    "grunt-peg": "^1.5.0",
     "grunt-release": "^0.7.0",
     "grunt-simple-mocha": "^0.4.0"
   }
